requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
PyYAML==6.0.1
streamlit==1.35.0
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
click==8.1.7
structlog==24.2.0
httpx==0.27.0        # async HTTP client for tests
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
rich==13.7.1          # beautiful CLI output
tabulate==0.9.0
